refactor: 그룹 사용자 기능 분리 (GroupUserService, GroupUserController) (#208)
